Create so-postinstall.sh
Security Onion supplies a non customised installer for Fleet. This script can be used by people who have an MDM or Munki to customise the installed so-launcher. 
Just add the required variables at the top, upload the .pkg that is generated inside Security Onion, and add this shell script as a postinstall script.

Comments and improvements are welcome.

More info here-
https://servicemax.com.au/tips/security-onion-automating-mac-host-install-for-fleet/